refactor: 도메인 계층과 인프라 계층으로 분리합니다.
