+++ src/main/java/lotto/domain/LottoAnalyzer.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class LottoAnalyzer {
     public static LottoReport analyze(final LottoTickets lottoTickets, final LottoTicket winning) {
